benchmarking-tool/cdk-stack/python-archive/requirements.txt
-e .

pip~=21.0.1
colorlog~=4.8.0
cryptography~=3.4.7
attrs~=20.3.0
python-dateutil~=2.8.1
publication~=0.0.3
constructs~=3.3.75
six~=1.15.0
setuptools~=54.2.0
docutils~=0.17
boto3~=1.17.45
apache-airflow
aws-cdk.aws-ec2==1.116.0
aws-cdk.aws-redshift==1.116.0
aws-cdk.aws-stepfunctions==1.116.0
aws-cdk.aws-stepfunctions-tasks==1.116.0